Split survey and collex helpers, remove survey_id, survey_name and collex_name from context
